Validations

+++ ConsoleApplication1/Program.cs
@@ -1,8 +1,5 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.Threading.Tasks;
 
 namespace ConsoleApplication1
 {
@@ -13,8 +10,8 @@ namespace ConsoleApplication1
             Console.WriteLine("Enter a number: ");
             int n = Convert.ToInt32(Console.ReadLine());
 
-            int baseTwo = NumericConversion.Transform.ToBinaryNumber(n);
-            int baseEight = NumericConversion.Transform.ToOctalNumber(n);
+            long baseTwo = NumericConversion.Transform.ToBinaryNumber(n);
+            long baseEight = NumericConversion.Transform.ToOctalNumber(n);
             string base16 = NumericConversion.Transform.ToHex(n);
 
             Console.WriteLine("Binary conversion: " + baseTwo);
